+++ src/Queue/src/Exception/RetryableExceptionInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spiral\Queue\Exception;
+
+use Spiral\Queue\RetryPolicyInterface;
+
+interface RetryableExceptionInterface
+{
+    public function isRetryable(): bool;
+
+    public function getRetryPolicy(): ?RetryPolicyInterface;
+}