+++ src/Queue/tests/Attribute/Stub/WithDefaultRetryPolicyAnnotation.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spiral\Tests\Queue\Attribute\Stub;
+
+use Spiral\Queue\Attribute\RetryPolicy;
+
+/**
+ * @RetryPolicy
+ */
+final class WithDefaultRetryPolicyAnnotation
+{
+}